skillfactory-project-main/chzsa/cars_management/serializers.py
from rest_framework import serializers

from .models import Car, Maintenance, Reclamation


class CarSerializer(serializers.ModelSerializer):
    class Meta:
        model = Car
        fields = '__all__'


class ServiceSerializer(serializers.ModelSerializer):
    class Meta:
        model = Maintenance
        fields = '__all__'


class ClaimSerializer(serializers.ModelSerializer):
    class Meta:
        model = Reclamation
        fields = '__all__'
