skillfactory-project-main/requirements.txt
Django==4.2.4
django-allauth==0.54.0
django-filter==23.2
djangorestframework==3.14.0